upd: board ごとに state を分けた
